backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
python-docx
firebase-admin
google-cloud-firestore
python-multipart
